replaced application.properties with application.yaml, add JWT authentication

+++ src/main/java/com/example/employeemanagement/exception/EntityNotFoundException.java
@@ -4,7 +4,7 @@ import com.example.employeemanagement.util.ErrorMessages;
 
 public class EntityNotFoundException extends RuntimeException{
 
-    public EntityNotFoundException(int id) {
+    public EntityNotFoundException(Long id) {
         super(String.format(ErrorMessages.EMPLOYEE_NOT_FOUND_MESSAGE, id));
     }
